chore: split coverage test
